PycharmProjects/CS673-Summer-2023-Team-6-Academix-main/server/requirements.txt
asgiref==3.7.2
certifi==2018.8.24
cffi==1.15.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.5.0
dj-rest-auth==4.0.1
Django==3.2.20
django-allauth==0.54.0
django-coreapi==2.3.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-coreapi-schema==0.1.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
gunicorn==19.7.1
idna==2.7
importlib-metadata==6.7.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
oauthlib==2.1.0
Pillow==9.5.0
pipreqs==0.4.13
pycparser==2.21
Pygments==2.2.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==6.0
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.3.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.24.2
whitenoise==3.3.1
yarg==0.1.9
zipp==3.15.0
